Added Task Detection
