+++ src/client/AM.PA.MonitoringTool/WindowsActivityTracker/Visualizations/DayFragmentationTimeline.cs
@@ -135,7 +135,7 @@ namespace WindowsActivityTracker.Visualizations
             html += "var " + activityTimeline + " = d3.timeline().width(" + TimelineZoomFactor + " * itemWidth).itemHeight(itemHeight)" + hover + mouseout + ";"; // .colors(colorScale).colorProperty('activity') // .stack()
             html += "var svg = d3.select('#" + activityTimeline + "').append('svg').attr('width', itemWidth).datum(data).call(" + activityTimeline + "); ";
           // html += "}; ";
-            html += " loadWindowTitleTimeline();}; "; // end #1, invoke DayWindowTitleList timeline
+            html += " loadTasksTimeline();}; "; // end #1, invoke DayWindowTitleList timeline
             html += "</script>";
 
             /////////////////////
